django/crm_project/requirements.txt
django 
djangorestframework 
psycopg2